requirements.txt
decorator==4.3.2
Django==2.1.5
django-crispy-forms==1.7.2
Pillow==5.4.1
pycodestyle==2.5.0
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
requests==2.21.0
users==1.0.dev0